mpl/gpu: Introduce gpu device abstraction
This includes MPL_gpu_device_handle_t as device abstraction and
an interface MPL_gpu_get_device_handle that provides the handle
for the device where a given memory address was allocated.

This is done in preparation to replace device id with device handle
in a future patch. For ZE API, it is more efficient to use device
handle instead of device id because many ZE functions accept device
handle as argument instead of device id. If we use device id,
we should translate it to the corresponding device handle before we
make ZE function calls, which involves a search loop.