ch4: add assertion to ensure number of vcis is within limit
To benefit from mutlti-vci and request pool, we need ensure a 1 to 1
mapping of vci and pool index. Otherwise, we won't be able to apply
granular locks per vci/pool neatly and resulting in effective
serialization.

The maximum of request pool is fixed at 64, which is a balance of handle
bits -- 64 vcis and 256*4096 requests should be enough for anyone --
quoting the famous rumor.