mpir: call unsafe handle functions with request objects
The request object need different default than the other handle
objects. In addition, we'll be shifting the VCI critical section to the
caller. It is critical for the fast patch to avoid incurring multiple
critical sections. If we make shift the critical section to caller,
caller is able to consolidate crtical sections.