ch4: add missing critical section around MPIR_Request_create
Now that we don't automatically add critical section inside
MPIR_Request_create (so we can consolidate critical sections in the
critical path), we'll need add necessary protection for some of the
request creations in ch4 that was outside critical section. Namely, they
are workq requests, prequests, and a pre-complted null_recv request.