ch4: register vci lock to request pool in ch4_init
To get the performance of multiple request pool, we need consolidate
per-vci lock with the per request pool lock, to avoid contention and
avoid cost of multiple critical sections in critical paths. To achieve
that, we essentially require a 1-to-1 vci-pool mapping and use the
per-vci lock to protect both the vci operations and request object
creation/free.