mpl: fixup for [bd07c1e62713]
The patch in [bd07c1e62713] does not work when the user passes
-DNDEBUG.  This version simply returns a NULL when the user asks for
more the PTRDIFF_MAX.  The system is not allowed to allocate more than
PTRDIFF_MAX anyway, and it is natural for the memory allocation
routines to return NULL when the requested allocation cannot be
satisfied.  So this patch might be more natural too.