mpl: fixup for [8089b2b4ff6c]
The patch in [8089b2b4ff6c] does not work when the user passes
-DNDEBUG.  This version simply returns a NULL when the user asks for
more the PTRDIFF_MAX.  The system is not allowed to allocate more than
PTRDIFF_MAX anyway, and it is natural for the memory allocation
routines to return NULL when the requested allocation cannot be
satisfied.  So this patch might be more natural too.