datatype: remove dataloop assumptions in max_contig_block calculation
The max_contig_block value should always be queried from typerep,
because different typerep backends could have different optimizations
for what they consider "contiguous".  MPICH assumes that its
computation of max_contig_blocks is the same as what the backend
datatype engine does.  That seems to work for dataloop, but is
incorrect for yaksa.

In this patch, we move the computation of the max_contig_blocks to
typerep, so typerep can have this MPICH assumption only for dataloop.
For yaksa, it would query the datatype engine to get this information
and to cache it.