typerep: convert typerep to a structure
MPICH tries to recalculate parts of the typerep engine (such as the
max_contig_blocks) internally because (1) querying typerep can be
expensive and (2) the typerep engine might not be ready by the time
these fields are important.  This is an incorrect design because, in
reality, MPICH does not know what optimizations the backend is using
to manage its data layout.  For example, in some cases, two blocks of
data that are contiguous might be treated as noncontiguous if it is
not easy to detect the data layout.  This calculation seems to match
what dataloop does, but does not match what yaksa does causing some
tests to fail with yaksa.

The right approach would be to always query the backend typerep engine
and cache the information inside MPICH.  The backend engine is ready
to be queried at datatype creation time for yaksa, but is only ready
at datatype commit time for dataloop.

In this patch, we refactor the code to pass the entire MPIR_Datatype
structure to typerep, so it can fill in more than just the typerep
backend engine handle.  This patch is almost entirely refactoring,
except the subarray/darray portions of typerep_commit, where the
original code was creating the dataloop for the temporary datatype
object, but assigning it to the main datatype object.  In that case,
we added a simple dup of the dataloop object to workaround this issue.