debuginfo: skip tracking builtin request in debugger queue
Pre-completed builtin request should be skipped since there won't be a
corresponding "forget" call.

Rationale: completed requests are immediately complted, so technically
it is not incorrect to not show up in debugger's radar. On the other
hand, in order to show up on debugger's radar, the implementation has to
be altered (instead of builtin request, heavy requests are created just
because of debugger) and the information revealed to debugger won't be
useful anyway.