datatype: redirect the old MPI_LB/UB API to the new API
We maintain a lot of extra code for the old MPI_LB/UB API usage.  This
is unnecessary.  It's easier to translate any usage of the old API to
the new API instead.  It adds a little bit of extra overhead, but: (1)
the overhead is at type creation time, and (2) users should really be
using the new API anyway.