ch4/ofi: add vni to MPIDI_OFI_REQUEST_CREATE and cleanup
Add the vni parameter so we can explicit control which vni to create
requests.

MPIDI_OFI_REQUEST_CREATE_CONDITIONAL and MPIDI_OFI_INIT_SIGNAL_REQUEST
are removed and usages expanded.  We need explicit awareness of how
the request is created. In particular, we may need to decide which vni
to create requests from and the choice of vni may be different for workq
path and regular path.