test/mpi/threads/perf: Pt2pt message-rate with a comm per thread
The multi-thread performance test help us identify performance issue,
especially due to avoidable thread contentions.

As a test rather than a benchmark, at places we prefer clarity
to reliability. Since there are too many factors for this test to pass,
we do not inlcude this folder in the test/mpi/threads/testlist.

Co-authored-by: Hui Zhou <hzhou321@anl.gov>