ch4/ofi: remove unused MPIDI_OFI_to_phys
This function is currently unused in master. With current design, a
`grank` only works with a process in `MPI_COMM_WORLD`, i.e. a newly
added dynamic process won't have a grank. Therefore, the use case for
this function is limited anyway. We'll need adapt the current av manager
with some universal grank system in order to use grank during runtime
cleanly.