ch4: add am-only configuration
This configuration forces all communication to only use the AM path.
Direct overrides of the netmod/shmmod for MPI functions will be
disabled, and the netmod/shmmod would be only used as byte transports.

It also removes the legacy-ofi (for OFI) and
ch4-netmod-ucx-args=am-only (for UCX) configure options, and adds
AM-fallback for SHM code too.