mpl/gpu: add addr offset in memory handle routines
CUDA IPC is always obtained based on the base address of the
memory region allocated by cuMalloc. Thus, we need internally query the
base address of a user-specified pointer and store the offset within the
handle. When open the IPC handle, the offset is added back to the
mapped pointer.