ch4: poll progress on each vci in MPIDI_Progress_test
We need enable polling progress on all vcis in order to make multiple
vci work.

Note: this is a quick patch to make multiple vcis work. While it should
perform the same when number of vcis is limited to 1, it will perform
very badly with more than one vci. A future patch will fix the
performance issue.