ch4/generic: using GenQ for AM recv on small unexpected message
For small unexpected messages, the pack buffer is allocated from a pool with
virtually unlimited amount (up to INT_MAX) of cells.

The pack buffer size (cell size) is 17 * 4KB to match the greater
between the POSIX iQueue cell size and the OFI eager threshold. The pack
buffer size is check against the eager buffer size of SHM and the pack
buffer size of NETMOD eager send at init time to make sure they are
large enough.

For large unexpected message, the pack buffer is still allocated with
MPL_gpu_malloc_host. This is because UCX do not differentiate small and
large messages, therefore the recv side could get really long
messages. This will be addresses when the pipeline AM protocol is added.