mpl/gavl: add remove list into gavl tree
gavl tree does not allow a node that is the subset of another. Therefore,
we remove all subset nodes of the input buffer during insertion. The removed
nodes are maintained at a separate list in the tree in order to ensure correctness
if any outstanding operations are still using these nodes. The removed nodes and
buffer objects are freed only when MPL_gavl_tree_delete is called.