ch4/ofi: refactor MPIDI_OFI_init_settings
Refactor MPIDI_OFI_init_global_settings into MPIDI_OFI_init_settings to
allow initiating a settings pointer. This will allow two separate
settings -- optimal_settings and minimal_settings -- to be used in
picking providers. Additional changes are made to not to rely on the
global settings, which are uninitialized at this point.

In addition to variable name changes, memset is added to initialize the
settings to zero since they are no longer global.