ch4/ofi: pick best provider
Instead of picking the first provider that match the default capability
set or drop to the first provider that match the minimal set, which
often omits providers that have capability set in between. For example,
`verbs;ofi_rxm` is often not picked simply because missing one
FI_ATOMICS flag, even though it is still far better than the minimal
set. This patch replace the original rejection based matching with a
score based matching, thus should be able to pick the best provider in a
single iteration.