ch4/ofi: fallback to MPIDIG for dynamic window
When dynamic window is created, we assume a base of 0 and disp_unit of
1. However, it is not clear whether calling `fi_mr_reg` on NULL address
is allowed and even when it is allowed, it is probably achieving
nothing. For correctness, we skip the memory registration for dynamic
window and fall back to active message for RMA operations.

TODO: add optimization to register each attached memory. We can
synchronize the mr_keys either via `symm_attach` hints (which require
collective MPI_Win_attach) or communicate key inside epoch
synchronization (which also require a hint that user won't detach memory
inside the epoch.