dataloop: add cast to supress ubsan warnings
Sometime bufp can be NULL and rel_off can be negative, and the result of
the pointer value become negative, causing undefined behavior. We
probably should differentiate the use case when the iovec is merely used
as displacements and when that is actually used as pointers. The former
should use a intptr_t type rather than void * to avoid undefined
behavior. Unfortunately, the usage in segment code is too complex to
sort, thus, this patch atopts the hacky solution of type cast.