ch4/ofi: post fi_trecv for closing message at cconnection time
The child process may go away and some provider (such as verbs;ofi_rxm)
will lose the disconnect message if the recv hasn't been posted yet.
While this seems to be a libfabric bug, since it is generally not a bad
idea to pre-post receiv buffer, we patch it by posting recv right at
connection time.