ch4/ofi: add vci protection for disconnect ofi calls
Previously disconnect is only called during finalize, where critical
sections are disabled. Now that we move part of the call to runtime,
we'll need add the vci protection to guard multi-thread usages. In
principle, we should allow disconnect during runtime, e.g. do real
disconnect in MPI_Comm_disconnect, therefore, it is a good idea to
assume we always need runtime critical sections. It is harmless if they
are called during finalize.