ch4/ofi: add wait after disconnect
When dynamic process disconnect and exit, some provider, e.g.
verbs;ofi_rxm, exhibit a bug that will lose the message if target
receive is not posted before exit. As a hack, wait a second to have
better chance of evading the issue. Since the process is disconnecting,
it should be harmless to wait a little.