ch4/nm: define RMA AM progress condition check hook
Through this hook, each netmod can report whether CH4 AM
progress is needed to complete the RMA operations issued
via netmod. If any AM based RMA may be issued, the netmod should
always return TRUE. Note that it has to be a global constant for each
window because a remote process may issue an AM to the other and require
the other to make AM progress.