ch4/ucx: wrap ucp_request_free to clear content
Ucx does not clear content of ucp_request before return to its internal
pool, potentially result in garbage content in the new request. We have
been setting its request handle to NULL to effectively clearing its
content, but that is too obscure to read and can easily miss cases -- we
did in ucx_progress.c and a few other places.

Let's use a wrapper to ensure we always clear content before returning
to ucx mpool.

Note: I believe the ucp_request_release usage in ucx_am.h is buggy, but
since we are going to revamp that code away, we patched it without fixing
in this commit.