ch4/ofi: get rid of magic number MPIDI_OFI_DT_SIZES
We defined MPIDI_OFI_DT_SIZES to set up atomics support for each
predefined datatype in ofi. Such a way can cause bug if we make change
in MPIR datatypes in future. This patch replaces it with
MPIR_DATATYPE_N_PREDEFINED. It allows us to also skip the index
initialization in ofi because we can get index from MPIR now.