qmpi: Add initial support for QMPI
To use this functionality, the tool needs to provide a function called
QMPI_Register_callbacks which returns a struct of function pointers that
the tool wishes to override (currently only send and recv are
supported). The tool must be compiled as a dynamically linked library
and at runtime, when running the application, the environment variable
QMPI_TOOL_LIST should include the name of the .so file containing the
tool. This can either be the name of the tool with the library in the
LD_LIBRARY_PATH, or a full path to the library file.

At runtime, MPICH will search for the tool with dlopen, call the
indicated function, and store the function pointers provided. When the
user calls an MPI function, MPICH will find the first entry in the
function pointer list and call that function (which will either be from
the tool or the MPI library itself). The tool is responsible for calling
the function QMPI_Get_function_table to get the full table of function
pointers and calling the next function in the table via the provided
input value of layer_id. Eventually, this will lead to the MPICH-level
function being called.

At finalization time, the tool is responsible for cleaning up its own
layer of the function table stack by freeing the memory and then calling
the next layer of the QMPI stack.