ch4/generic: implement RDMA_READ protocol
The RDMA_READ protocol is initiated by sending the RDMA_READ_REQ
message. This messsage is handled by the netmod level RDMA_READ send
function which takes care the memory registration and transmissing
memory registration info.

On the receiver side, when the RDMA_READ_REQ is received and matched, it
calls MPIDIG_do_rdma_read_recv which calls corresponding NM/SHM function
to attemp initiating the RDMA read. Note, once the RDMA read is
initiated (may not completed), the function will return and finishes the
execution of target callback. Therefore, we are not sending
RDMA_READ_ACK inside the callback, nor do we complete the receive
request. The NM/SHM will internally track the completion of RDMA_READ
then issue the ACK and complete the request.

When the ACK arrives at the sender, the sender will call NM/SHM
interface for unregister the send buffer.

If the NM/SHM decide to reject RDMA_READ protocol, it will need to call
MPIDIG_do_rdma_read_nak to send the NAK message. The sender will
fallback to PIPELINE protocol. In this case, the pipeline will start
from sending PIPELINE SEG directly as the message has already been
matched on the receiver side.