ch4/wait: get vci from real_request for progress
Workq requests are pseudo requests. We need get vci from its real
request just as persistent requests.

For persistent requests nested with workq requests, we need go one
level deep to get real request. And this is also a good place to update
the persistent request's cc_ptr to the real real cc.

NOTE: the persistent cases is rather tricky to get it right, thus the
extra in-code comments.