ch4/ofi: Fix the underflow issue in huge message protocol
Huge message protocol in MPICH was unable to handle underflow when the sender sent
a message of size exactly equal to `max_msg_size`. Since the sender would send this message
using a `tsend` but the receiver would expect to get a huge message header from the sender.
This issue has been fixed by following the huge message code path for sending messages of size
equal to `max_msg_size`.