ch4: refactor rma window sync progress macros
The window synchronization progress are only concerned with progress
within the window, and may only progress for the vci that is being used
for the window. We use a much simplified version of progress function
here to be used inside a per-vci critical section, thus avoid repeatedly
exiting and entering locks, and avoid unnecessarily polling global progress.

Poll global progress in MPIDI_progress_test_vci.

Some time the application may require progress in another vci to unblock
the progress in a different vci.

An example is test/mpi/rma/mutex_bench_shm_ordered, where the ofi
lightweight send needs progress but not triggered due to the window
being in another vci.