ch4/ofi: move rma_sep and rma_stx_ctx to per-vni
Because both rma_sep and rma_stx_ctx need bind to the specific domain,
av, and cq, we need move them to per-vni to work when multiple vni is
enabled.

For the case when we use scalable endpoints to implement vni, both of
these can be shared (just as domain and av). We can easily manage them
by simply aliasing them across all vnis (just as domain and av). This is
left as TODO (when we decide it's worthwhile).

Moved the init of rma shared context code from init to windown init.
This is consistent with how rma_sep is coded.