ch4/ofi: remove MPIDI_OFI_av_insert
When this wrapper was initially introduced, it has a couple #ifdefs --
now removed -- and it doesn't have the current vni_local/vni_remote
complications. Now that the #ifdefs are removed, and the actual usage
had more variations that made this wrapper function no longer useful.

Not only that, it as currently written, is incorrect. It works because
we only used it for the vni=0 case.

This commit removes this wrapper. If there will be future need -- for
example, the av AV struct grows sufficent sophistication that warrants
some wrapper, we can re-introduce it with more correctly tuned
interface.