misc: use MAX_ALIGNMENT rather than hardcode it
Hard coding the alignment requirement is a bit undesirable. Why 8 byte
alignment seems work for most of the case, it doesn't prevent on certain
system or applications that user may need 16-byte or 4-byte alignment.
This commit replace the magic number with MAX_ALIGNMENT, which is set in
mpichconf.h by configure.