ch4/posix: allocate POSIX request cache with local size
The posix request cache contains the on-going recv request for different
source. The size of the array should be the number of local processes
rather than the entire comm world.

To make this change, it also needed to change the eager interface to set
src_local_rank instead of src_grank on receiving.