mpl: limit scope on macos .local workaround
This was added in 434dca66ff9b6f496b89062b41af8d360ed813cc.
Assuming that "localhost" is equivalent to any name ending in .local
is unwarranted. RFC6762 reserves the ".local" suffix for MulticastDNS use,
but it is also used in other context for "local network" addresses and such.
So let's at least limit the scope to not hurt other systems.