ch4/ofi: move MPIR_CVAR_CH4_EAGER_MAX_MSG_SIZE back to ofi
This cvar was desgined controls the message size at which OFI native
path switches from eager to rendezvous mode. It was not intent to affect
the AM path eager limit.

This commit move it back to OFI and rename it to
MPIR_CVAR_CH4_OFI_EAGER_MAX_MSG_SIZE. This commit also restore the code
that initialize MPIDI_OFI_global.max_msg_size with the CVAR.