ch4/generic: add utility function for tracking pipeline
Setup internal counters for how much data need to be send, current
offset. It also tracks how many segment are on the fly.

Callback is changed to do MPIDIG_recv_type_init for long message in the
send callback rather than in data_seg callback as the later won't be
able to correct init the recv for message size in pipeline