ch4/ofi: implement am pipeline send
The pipeline send splits long message into segments that are smaller
than eager threshold. The send operation will try to send the am_hdr and
the first segment, then enqueue the rest send to the deferred queue. The
progress will be responsible for issuing the rest segment.

Created new amtype for pipeline so that the receiver can use a different
path to handle the pipeline recv. When the first segment arrives, the
receiver will call the target callback function to get a recv request.
This request is store in a cache if more data is expected. The
subsequent segment will only need to get the request from the cache
before recving the data.