ch4/generic: split async structure for send and recv.
Send and recv async structure should not share a union becuase for AM
get accumulate, the same request is used for sending and receiving. The
receiving operation would overwrite the send async data.

Here is the case of how it happens, the ACK of GET_ACC may come back
before the sending is locally completed. This is possible if the send is
pipelined, the data on segments may have arrived on target process but
those segments are still waiting for local completion. Since the target
process already have the data, it will proceed with accumulate and send
back the ACK data. When using OFI netmod, the cq_read may return the
event for the incoming ACK data before the local completion of pipeline
send, therefore triggers the GET_ACC_ACK callback to init the request
for receiving by setting the recv async structure which overwrite the
send async structure that is still needed for the local completion of
sends.