ch3/config: update ofi library configure
Simplify a little since now PAC_CHECK_HEADER_LIB_FATAL is now more
sophiscated. In particular, --with-ofi (without setting path) will not
add "yes" to the include and library paths now.

Note, we preserved the option to use --with-ofi, which is different from
ch4 configure, which uses --with-libfabric. There is nothing different
between ch3 ofi and ch4 ofi, so we could unify the behavior and in
addition allow ch3 ofi to use the embedded libfabric as well. The is
left to the future.