ch4/rma: use mpir op shortname to simplify info which_accumulate_ops
RMA program may specify which_accumulate_ops=min,max... to define the
used atomics in order to enable network hw atomics when available.
Note that cswap is an RMA atomic but not reduce op, thus we need
specially handle it as OP_NULL. The value was statically parsed.
This commit uses MPIR Op shortname routines to simplify it.