ch4/ofi: force per-window native RMA if user prefers
When user sets hint rma_issue_mode=native, we check whether it is
possible to force native path for all RMA. We consider three kinds of
limitations:
1. Hardware limit: e.g., an atomic operation cannot be supported by network
2. Implementation limit: e.g., we don't have native version for noncontig
   accumulates and no native support for cross-mem op in dynamic win
3. Performance limit: e.g., we choose to fallback to AM for sparse
   put/get

The former two are mandatory for correctness, thus we need check whether
these limits exist at window creation and enable force_native only when
possible; the third is optimization, thus we can choose to ignore it
when user requires.

This hint is for advanced users (e.g., a upper library of MPICH). If
need to gain performance from limit-3, the user should not set this
hint.