ch4/ofi: remove MPIDIG_wait_am_acc in native atomics
MPIDIG_wait_am_acc is insufficient to guarantee the completion of active
message atomics since there is no way of accounting for remote
initiating active messages that is targeting the process. Since we
essentially enforced the native-only mode, in which no active messages
should happen, we can remove this code.