maint: add maint/mpi_standard_api.txt
This is the base file for generating MPI bindings, including the C
binding.

The file is scraped from the official mpi-standard repository with a
script. Any deviations or customizations, such as descriptions, notes,
implementations, etc. should be maintained in separate files.

Currently this is based on mpi-standard mpi-4-rc branch at:
6d9cdda 11/30 15:05 Merge pull request #343 from wgropp/captions