ch4/netmod: remove size check for native PUT and GET
The MPI standard does not require the data sizes represented by the
origin and target datatype signatures to be equal. It only requires
origin data to fit in target buffer in MPI_PUT, and target data to fit
in origin buffer in MPI_GET. This check is overly restrictive and we can
save it by not checking anything.

For MPI_GET, this means the null_op check should be on if
target_bytes zero rather than origin_bytes.

This commit also cleanup unused origin_bytes and result_bytes variables.