ch4/generic: cleanup and clarify data_sz related in RMA operations.
For put and get, the target_count in message header for non-predefined
datatype is set to target_data_sz now as the standard allows it to be
different from origin_data_sz. We should not assume the origin_data_sz
is equal to it.

Data sizes involved in RMA are clearly named with origin, target, result
prefixes. All data size variable touched is defined with MPI_Aint now.