ch4/posix: refactor POSIX EAGER interface
Add buffer signature and am header to the send interface so it can
handle non-contiguous send buffer.

The data_sz_left and offset are passed explicitly so that the pipelining
is still implemented at the POSIX level. Therefore not requiring EAGER
module to handle long message. The eager module would only send the data
up to its limit and return how much data is sent for POSIX to track
progress.

The request object is not passed by design as in the case of sending
header only message, no request is provided.