ch4: cleanup MPID_Alloc_mem code
The current MPID_Alloc_mem code is completely bogus.

1. All allocations are immediately forwarded to the netmod.  Shmmods
are completely ignored.

2. If the user requests for bound HBM or DDR memory, this is
completely ignored in the primary path.  Only if the netmod fallsback
to the generic layer are these requests even considered.

3. It maintains a linked list of memory allocations and always
searches through them for frees.  While this function is not
performance critical, maintaining an O(N) search-time data structure
for this seems silly when one could have just as easily used a hashmap
to make the lookup faster.

4. The linked list insertions and deletions are not protected by the
per-object lock.  This works fine for the global lock, but would be
broken for per-object locks.