init: add init_lock and init_counter
It is allowed to have MPI_Session_init to run concurrently with
MPI_Init or another MPI_Session_init in a different thread, so we need
lock to control race conditions. We use MPL_initlock to avoid the
initialization issue of normal mutexes.

We also add init_counter so to track multiple init and finalize.