init: add lock and counter to control multiple init and finalize
It is allowed to have MPI_Session_init to run concurrently with
MPI_Init or another MPI_Session_init in a different thread, so we need
lock to control race conditions. We use spin lock to avoid the
initialization issue of normal mutexes.

This lock can be used for thread-safety of other routines that are
allowed to be accessed outside init/finalize.

We also add init_counter so to track multiple init and finalize.