init: add and use MPIR_Errutil_is_initialized
Rather than verbosely compare the mpich_state, which is semantically
obscure, wrap it in a function to make its semantics clear.

Use MPIR_Errutil_is_initialized in MPIR_Err_return_comm.  The usage is
for the same purpose as other usage of MPIR_Errutil_is_initialized,
that if the error utitlities are not ready, we'll have no choice but
to abort here.