mpi/err: revamp the error code/class data structures
We were originally using a single counter to keep track of the next
available error code/class.  This was fine for MPI-3.1, where users
could only add error codes/classes/strings, but could not delete them.
But in MPI-4, we have routines to delete them as well.  In such cases,
a single counter will not work.

This patch converts the data structure used to a collection of a
counter, a free doubly-linked-list, and a used hashmap.  When a class
or code is created, we first check the free list.  If none is
available, we'll increment the integer and add it to the hashmap.
When a class or code is freed, we remove it from the hashmap and add
it to the linked list.  This is still an O(1) creation/deletion data
structure.

Note that these functions are not thread-safe (they were not
thread-safe even before this patch), so the expectation is that
thread-safety is handled elsewhere.