maint: add gen_binding_c.py
This is the script to be used to generate C bindings based on
maint/mpi_standard_api.txt and additional custom files.

The script puts bulk of the routines in python local package module,
local_python. This should make the code easier to maintain and potential
for future script reuse.