ch4: adding recv_ready field for request
This field is used for indicating if the request is ready for receiving
data. The default is set to true as in most cases (all RMA messages and
short SEND message) it is ready for data. The term "ready" here is equal
to "recv buffer info has been filled out in request so any recv_copy
operation will be moving data into its proper destination".

The case where this would be set to false if when the message is
unexpected and the transport is requesting for async recv. Setting this
field (and having the transport checking it before copying data)
prevents the transport form copying data when buffer is not ready, but
allowing the transport to save any necessary information into the
transport region of the request.