darray (main path) needs a resize in the block()/cyclic() subtype creation
When Type_create_darray() uses the block()/cyclic() function to
construct a subtype for whatever dimension it's processing, it
needs to resize the resulting type before it goes into the type
construction as the input for processing the next dimension.

Gist for a testcase:
https://gist.github.com/markalle/940de93d64fd779e304ee124855b8e6a

The darray_bug_romio.c testcase creates a darray using
* 4 ranks in a 2x2 grid
* looking at the type made for rank 0
* inner dimension: 4 ints distributed block over 2 ranks with 2 items each
* outer dimension: 6 of the above distributed cyclic over 2 ranks with 2 items each

The type created by processing the first dimension should look like this
[ x x . . ]

And then distributing 6 of the above for the outer dimension would be
[ x x . . ]
[ x x . . ]
[ . . . . ]
[ . . . . ]
[ x x . . ]
[ x x . . ]

But if we remove this PR's resize at the bottom of MPIR_Type_block()
for example, the first dimension is just
[ x x ]
And then distributing those for the second dimension becomes

[ x x x x ]
[ . . . . ]
[ . . . . ]
[ . . . . ]
[ x x x x ]
[ . . . . ]

Going to the MPI standard to justify why the first layout is right,
it's where the definiton of the cyclic() function has a ub_marker
of gsize*ex, eg 4 ints for that first type.

I did notice datatype_impl.c has a resize already in its cyclic() function,
but I wasn't sure if that was trying to handle this case or not.  If so it's
not handling the case right as "ex" in the standard isn't the orig_extent,
it's the extent of oldtype that came from the previous dimension.  Eg in
the outer dimension above gsize=6 and ex=16 when setting the size for the
final type.

Signed-off-by: Mark Allen <markalle@us.ibm.com>