make ADIOI_GEN_WriteStrided not step on itself
The ADIOI_GEN_WriteStrided funcion uses data sieving on non-contiguous
types. That is, if it wants to write data at locations
    [W...X...Y...Z...]
it reads the whole buffer
    [dddddddddddddddd]
changes the locations it wants to write to
    [WdddXdddYdddZddd]
then writes the whole thing back.  It uses locks to make this safe, but
the problem is this only protects against other parts of the product that
are using locks.  And without this PR a peer who is simultaneously making
a simple non-contiguous write wouldn't have locked.

In the fully general case I'm not positive this commit makes WriteStrided
safe as I'm not confident that all WriteContig calls are protected by locks.
But in practice I couldn't come up with a testcase that failed after putting
in this fix.  It turned out that a wide variety of calls end up back at
the location where I added the lock, so a lot of code is made safe by this
checkin.

A testcase to demonstrate the original problem is here:
    https://gist.github.com/markalle/d7da240c19e57f095c5d1b13240dae24
% mpicc -o x romio_write_timing.c
% mpirun -np 4 ./x
Note: you need to use a filesystem that uses ADIOI_GEN_WriteStrided to
hit the bug.  I was using GPFS.

This commit adds detection at File_set_view() time to see if any of
the ranks are using a non-contiguous view, and uses that to key off of
later in GEN_WriteStrided() to make a rank lock in that case even if
its personal view is contiguous.

Part of this commit is in ad_io_coll.c where a temporary view is created.
I don't have a testcase that goes through that path.  I wasn't sure just
from reading the code whether all calls into ADIOI_IOFiletype() were
collectives, so I didn't want to do an Allreduce there.  So instead I
went a level up and attached an attribute to a temporary datatype that
was created by its caller.  That may have been unnecessary, but that's
how I made sure I was only every running my Allreduce from a location that
I was sure was indeed collective.

Signed-off-by: Mark Allen <markalle@us.ibm.com>