ch4: consistently use _p suffix for output pointer parameters
It is confusing to use the same name for output parameter, for example,
win is MPIR_Win * in some place but MPIR_Win ** in window creation
function. Since we need access the fields of the newly created objects
inside the creation function, the same name different types cause visual
confusions. This commit consistently uses _p for output pointers at the
API level.