ch4: remove create_intercomm_from_lpids from ch4 api
This api function is never used or implemented. Its name suggests it is
to create an intercommunicator, then it seems the parameter type for
newcomm_ptr is incorrect as it should be MPIR_Comm ** rather than
MPIR_Comm *.

Remove rather than fixing it. We always can add it back if it is needed
in the future.