ch4: use MPI_Aint for count parameter instead of MPI_Count
MPI_Count is identical as MPI_Aint on 64-bit system. But on 32-bit,
MPI_Count is 64-bit and less efficient while MPI_Aint is 32-bit. Since
the MPID API doesn't use MPI_Count, we should use MPI_Aint for
consitency.

Note that count-3 is `int` and they are used for collectives. Because
currently MPIR_Xxx_impl functions for collectives are using `int`, so we
need keep it compatible. As part of big count development, we will
update all internal count types to `MPI_Aint`, so `count-3` will be
changed to `count` as MPI_Aint soon.