python: rma userbuffer validation always check datatype
In pt2pt or coll, there are legitimate use cases for zero-sized
messages, in which case, the datatype is often invalid such as
MPI_DATATYPE_NULL, so we need skip the validation when count is 0.
But it doesn't make sense to have zero sized message in RMA. So we
should always validata datatype since it probably is a mistake for user
to pass in invalid datatype.