python: allow custom cleanup code in the c binding
We can add custom handle_ptr routine for the case when we want to skip
automatic handle validations. With handler array, we'll also need custom
clean up code to free the pointers array that is allocated in the custom
hendle_ptr routine. This is needed for MPI_Comm_spawn_multiple.