python: fix _comm_from_request
In MPI_{Test,Wait}, we use the comm_ptr from the request to be used for
error handling in fn_fail. Add a comment where we save the comm_ptr
since this is not obvious.

We only do this for `MPI_Test` and `MPI_Wait`. It is not necessarily
clear which communicator to use when we have an array of request such as
in `MPI_Testall`, so we are just using NULL comm in that case.