ch4/startall: refactor startall for partitioned req
Startall will accept both persistent req and partitioned req. This
commit refactors the code thus we can add more kinds.

Handling of persistent req is moved to MPIDI_prequest_start and we check
the kind of request in MPID_Startall. Err is reported if it is not persistent
req kind.