ch4: add netmod/shm api init_local and init_world
Some of the local functionality such as datatype commit will depend on
the initialization of netmod such as ucx. Add and split `mpi_init_hook`
into `init_local` and `init_world` to initialize the local part of
netmod.

We'll start with most of the current init in `init_world` to minimize
the impact to existing code, and only move necessary parts to
`init_local` to allow necessary local functionality.