test: add mult-thread session test
Note: this test is slightly modified from init/session.c. It removed the
initial info hint because at the time of this commit, the info usage
outside init/finalize is not merged yet and the test of info usage is
not the scope of this test.

This test will test both mpi://WORLD and mpi://SELF