make ADIOI_GEN_WriteStrided not step on itself
The ADIOI_GEN_WriteStrided funcion uses data sieving on non-contiguous
types. That is, if it wants to write data at locations
    [W...X...Y...Z...]
it reads the whole buffer
    [dddddddddddddddd]
changes the locations it wants to write to
    [WdddXdddYdddZddd]
then writes the whole thing back. It uses locks to make this safe, but
the problem is this only protects against other parts of the product that
are using locks. And without this PR a peer who is simultaneously making
a simple non-contiguous write wouldn't have locked.

A testcase to demonstrate the original problem is here:
https://gist.github.com/markalle/d7da240c19e57f095c5d1b13240dae24

% mpicc -o x romio_write_timing.c
% mpirun -np 4 ./x

Note: you need to use a filesystem that uses ADIOI_GEN_WriteStrided to
hit the bug. I was using GPFS.

This commit is pulled from wkliao after discussing where to put the
new lock.  It adds locks to contiguous writes in independent write
functions when data sieving write is not disabled

Signed-off-by: Mark Allen <markalle@us.ibm.com>