ch4/ofi: generate collective mr_key in win init
For dynamic window, when collective attach is enabled, we still use local mr key
for each buffer registration during win attach. For other types of windows, when
MPIDI_OFI_ENABLE_MR_PROV_KEY is OFF, MPICH needs to guarantee the processes
in a window use the same mr_key. Previous code does not check whether the mr_key
is the same or not which causes the bug.  This patch fixes this issue.