mpir/objects: do not add finalization hook if obj_kind is MPIR_INFO
As of MPI-4, MPI_Info functions can be called before MPI_Init() or after
MPI_Finalize(), so some MPI_Info objects may not be freed when MPI_Finalize() is
called.  This patch adds a branch in MPIR_Handle_obj_alloc_unsafe() to skip
finalization hook if obj_kind is MPIR_INFO.

Note that this is check only when objmem is not initialized, so it does not
affect the performance.