ch4/am: define am-specific struct for partitioned req
To support basic am-based partitioned pt2pt, we need store additional
info in the partitioned request. The am-based sycn will rely on an
atomic status on both sender and receiver. It is atomic because status
changes may be performed by different threads (steps (2) and (3) below).

On sender, the status change follows three steps:
  (1) psend_init: set 0 [UNSET];
  (2) start: inc 1;
  (3) received receiver's CTS: inc 1.
  Steps (2) and (3) may be out-of-order, but we need only wait on the
  [CTS == 2] status when ready to transfer data. After completed a req,
  the status is reset to 0 [INACTIVE].

  For second-round start-complete or later, it needs both steps (2) and
  (3) to reach [CST] again.

On receiver, the first-round status change follows:
  (1) precv_init: set 0 [UNSET];
  (2) start: inc 1;
  (3) rreq matched: inc 1.
  Similarly, steps (2) and (3) may be out-of-order. It notifies the
  sender CTS only when the local status becomes [CTS == 2].
  When completed a req, the status is reset to 1 [INACTIVE].

  For second-round start-complete or latter, it needs only step (2) to
  reach [CTS] again.