fortran: hardcode MPI_STATUS_SIZE
The configure check of MPI_STATUS_SIZE was silly that it hard code the
struct as 5 ints and to check whether the compiler agrees it is 5 ints
-- of course. Just hard code the result to keep it simple.

Note: the hardcoded size need be synchronized with the MPI_Status
definition in mpi.h. This dependency is not new however. We always
needed to patch both configure and buildiface whenever we change the
definition of MPI_Status.