fixes related to different build environments
I had a build on a PPC64 RH7.5 system that failed saying
aligned_alloc() needed declared.  What happened is _GNU_SOURCE
was used during configure (comes from mpl/include/config.h.in)
so when automake queries if aligned_alloc() is declared it
detects that it is.  But the include order used in various
.c/.h doesn't guarantee that mpl.h is included before stdlib.h
and the declaration only appears if _GNU_SOURCE comes before
stdlib.h

Signed-off-by: Mark Allen <markalle@us.ibm.com>